+++ src/main/java/com/syaru/advancedquantumengineering/block/ModifiedDataEntanglerBlock.java
@@ -6,13 +6,12 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.pedroksl.advanced_ae.common.blocks.AAEAbstractCraftingUnitBlock;
-import net.pedroksl.advanced_ae.common.blocks.AAECraftingUnitBlock;
 import net.pedroksl.advanced_ae.common.blocks.AAECraftingUnitType;
 import org.jetbrains.annotations.Nullable;
 
 public class ModifiedDataEntanglerBlock extends AAEAbstractCraftingUnitBlock<ModifiedDataEntanglerBlockEntity> {
     public ModifiedDataEntanglerBlock() {
-        super(AAECraftingUnitBlock.getProps(AAECraftingUnitType.STORAGE_MULTIPLIER, false), AAECraftingUnitType.STORAGE_MULTIPLIER);
+        super(metalProps(), AAECraftingUnitType.STORAGE_MULTIPLIER);
     }
 
     @Nullable