+++ src/main/java/com/syaru/advancedquantumengineering/block/ModifiedQuantumAcceleratorBlock.java
@@ -6,13 +6,12 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.pedroksl.advanced_ae.common.blocks.AAEAbstractCraftingUnitBlock;
-import net.pedroksl.advanced_ae.common.blocks.AAECraftingUnitBlock;
 import net.pedroksl.advanced_ae.common.blocks.AAECraftingUnitType;
 import org.jetbrains.annotations.Nullable;
 
 public class ModifiedQuantumAcceleratorBlock extends AAEAbstractCraftingUnitBlock<ModifiedQuantumAcceleratorBlockEntity> {
     public ModifiedQuantumAcceleratorBlock() {
-        super(AAECraftingUnitBlock.getProps(AAECraftingUnitType.QUANTUM_ACCELERATOR, false), AAECraftingUnitType.QUANTUM_ACCELERATOR);
+        super(metalProps(), AAECraftingUnitType.QUANTUM_ACCELERATOR);
     }
 
     @Nullable