+++ src/main/java/com/syaru/advancedquantumengineering/block/ModifiedQuantumMultiThreaderBlock.java
@@ -6,13 +6,12 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.pedroksl.advanced_ae.common.blocks.AAEAbstractCraftingUnitBlock;
-import net.pedroksl.advanced_ae.common.blocks.AAECraftingUnitBlock;
 import net.pedroksl.advanced_ae.common.blocks.AAECraftingUnitType;
 import org.jetbrains.annotations.Nullable;
 
 public class ModifiedQuantumMultiThreaderBlock extends AAEAbstractCraftingUnitBlock<ModifiedQuantumMultiThreaderBlockEntity> {
     public ModifiedQuantumMultiThreaderBlock() {
-        super(AAECraftingUnitBlock.getProps(AAECraftingUnitType.MULTI_THREADER, false), AAECraftingUnitType.MULTI_THREADER);
+        super(metalProps(), AAECraftingUnitType.MULTI_THREADER);
     }
 
     @Nullable