+++ src/main/java/com/syaru/advancedquantumengineering/item/AQEUnitBlockItem.java
@@ -10,9 +10,7 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Rarity;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
-import org.jetbrains.annotations.Nullable;
 
 public class AQEUnitBlockItem extends BlockItem {
     public AQEUnitBlockItem(Block block) {
@@ -25,7 +23,8 @@ public class AQEUnitBlockItem extends BlockItem {
     }
 
     @Override
-    public void appendHoverText(ItemStack stack, @Nullable Level level, List<Component> tooltip, TooltipFlag flag) {
+    public void appendHoverText(
+            ItemStack stack, Item.TooltipContext context, List<Component> tooltip, TooltipFlag flag) {
         Block block = getBlock();
         if (block == AQEBlocks.MODIFIED_QUANTUM_STORAGE.get()) {
             tooltip.add(Component.translatable(