+++ src/main/java/com/syaru/advancedquantumengineering/item/BigIntegerQuantumCoreItem.java
@@ -10,9 +10,7 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Rarity;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
-import org.jetbrains.annotations.Nullable;
 
 public final class BigIntegerQuantumCoreItem extends BlockItem {
     public BigIntegerQuantumCoreItem(Block block) {
@@ -27,7 +25,7 @@ public final class BigIntegerQuantumCoreItem extends BlockItem {
     @Override
     public void appendHoverText(
             ItemStack stack,
-            @Nullable Level level,
+            Item.TooltipContext context,
             List<Component> tooltip,
             TooltipFlag flag) {
         tooltip.add(Component.translatable(