+++ src/main/java/com/syaru/advancedquantumengineering/registry/AQECreativeTabs.java
@@ -5,15 +5,15 @@ import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.RegistryObject;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.neoforge.registries.DeferredRegister;
 
 public final class AQECreativeTabs {
     private static final DeferredRegister<CreativeModeTab> TABS =
             DeferredRegister.create(Registries.CREATIVE_MODE_TAB, AdvancedQuantumEngineering.MODID);
 
-    public static final RegistryObject<CreativeModeTab> MAIN = TABS.register("main",
+    public static final DeferredHolder<CreativeModeTab, CreativeModeTab> MAIN = TABS.register("main",
             () -> CreativeModeTab.builder()
                     .title(Component.translatable("itemGroup.advanced_quantum_engineering"))
                     .icon(() -> new ItemStack(AQEItems.MODIFIED_QUANTUM_CORE.get()))